+++ docs/changelog.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- **⚠️Experimental**: Bringing back the support for React output target.
+- **⚠️Experimental**: Adding example React app with Vite.
+
 - [Documentation](guide/angular.md) for using web components with Angular's standalone bootstrapping
 - `six-header-dropdown`: Added `filter` and `filterPlaceholder` properties.
 - `six-header-menu-button`: Added `caret`, `disabled`, `loading`, `submit` and `reset` properties.